# xwSystem0611/xwRAG/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "RAG知识库系统 API",
    "description": "基于Java的RAG知识库系统API，支持文档导入、智能切片、混合检索",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "本地开发环境"
    }
  ],
  "tags": [
    {
      "name": "健康检查",
      "description": "系统健康状态检查"
    },
    {
      "name": "文档接口",
      "description": "文档导入相关接口"
    },
    {
      "name": "检索接口",
      "description": "文档检索相关接口"
    }
  ],
  "paths": {
    "/api/v1/ping": {
      "get": {
        "tags": ["健康检查"],
        "summary": "Ping存活检查",
        "operationId": "ping",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseString"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/health": {
      "get": {
        "tags": ["健康检查"],
        "summary": "详细健康检查",
        "description": "检查系统各组件状态（Elasticsearch、Embedding、数据库、OCR）",
        "operationId": "health",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseHealth"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/upload": {
      "post": {
        "tags": ["文档接口"],
        "summary": "上传DOCX文件",
        "description": "上传DOCX文件并自动解析、切片、向量化、存储",
        "operationId": "uploadDocument",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["file"],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "DOCX文件"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "上传成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDocumentUpload"
                }
              }
            }
          },
          "400": {
            "description": "请求错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/import/mysql": {
      "post": {
        "tags": ["文档接口"],
        "summary": "从MySQL导入数据",
        "description": "支持全量导入(all)、按类型导入(bytype)、按ID列表导入(byids)",
        "operationId": "importFromMySQL",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MySQLImportRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "导入成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseMySQLImport"
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/search": {
      "post": {
        "tags": ["检索接口"],
        "summary": "检索查询",
        "description": "支持三种检索模式：hybrid(混合)、bm25(关键词)、vector(向量)",
        "operationId": "search",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchQueryRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "检索成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseSearch"
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/search/bm25": {
      "get": {
        "tags": ["检索接口"],
        "summary": "BM25关键词检索",
        "description": "基于关键词的全文检索",
        "operationId": "bm25Search",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": true,
            "description": "查询关键词",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "topK",
            "in": "query",
            "required": false,
            "description": "返回结果数量",
            "schema": {
              "type": "integer",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "检索成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseSearch"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/search/vector": {
      "get": {
        "tags": ["检索接口"],
        "summary": "向量语义检索",
        "description": "基于语义相似度的向量检索",
        "operationId": "vectorSearch",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": true,
            "description": "查询文本",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "topK",
            "in": "query",
            "required": false,
            "description": "返回结果数量",
            "schema": {
              "type": "integer",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "检索成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseSearch"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ApiResponseString": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "string"
          },
          "timestamp": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "ApiResponseHealth": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object",
            "properties": {
              "status": {
                "type": "string"
              },
              "timestamp": {
                "type": "string"
              },
              "components": {
                "type": "object",
                "properties": {
                  "elasticsearch": {
                    "type": "object",
                    "properties": {
                      "status": {
                        "type": "string"
                      },
                      "note": {
                        "type": "string"
                      }
                    }
                  },
                  "embedding": {
                    "type": "object",
                    "properties": {
                      "status": {
                        "type": "string"
                      },
                      "note": {
                        "type": "string"
                      }
                    }
                  },
                  "database": {
                    "type": "object",
                    "properties": {
                      "status": {
                        "type": "string"
                      },
                      "note": {
                        "type": "string"
                      }
                    }
                  },
                  "ocr": {
                    "type": "object",
                    "properties": {
                      "status": {
                        "type": "string"
                      },
                      "note": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "timestamp": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "ApiResponseDocumentUpload": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/DocumentUploadResponse"
          },
          "timestamp": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "DocumentUploadResponse": {
        "type": "object",
        "properties": {
          "documentId": {
            "type": "string",
            "description": "文档ID"
          },
          "filename": {
            "type": "string",
            "description": "文件名"
          },
          "chunks": {
            "type": "integer",
            "description": "切片数量"
          },
          "status": {
            "type": "string",
            "description": "状态"
          },
          "elapsedTime": {
            "type": "integer",
            "format": "int64",
            "description": "处理耗时(毫秒)"
          },
          "metadata": {
            "type": "object",
            "description": "元数据"
          }
        }
      },
      "MySQLImportRequest": {
        "type": "object",
        "required": ["type"],
        "properties": {
          "type": {
            "type": "string",
            "description": "导入类型: all/bytype/byids",
            "enum": ["all", "bytype", "byids"]
          },
          "typeId": {
            "type": "integer",
            "description": "类型ID(type=bytype时必填)"
          },
          "ids": {
            "type": "array",
            "description": "ID列表(type=byids时必填)",
            "items": {
              "type": "integer"
            }
          }
        }
      },
      "ApiResponseMySQLImport": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object",
            "properties": {
              "totalRecords": {
                "type": "integer",
                "description": "总记录数"
              },
              "successCount": {
                "type": "integer",
                "description": "成功数量"
              },
              "failCount": {
                "type": "integer",
                "description": "失败数量"
              },
              "elapsedTime": {
                "type": "integer",
                "format": "int64",
                "description": "耗时(毫秒)"
              }
            }
          },
          "timestamp": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "SearchQueryRequest": {
        "type": "object",
        "required": ["query"],
        "properties": {
          "query": {
            "type": "string",
            "description": "查询文本"
          },
          "topK": {
            "type": "integer",
            "description": "返回结果数量",
            "default": 10
          },
          "searchType": {
            "type": "string",
            "description": "检索类型",
            "enum": ["hybrid", "bm25", "vector"],
            "default": "hybrid"
          },
          "bm25Weight": {
            "type": "number",
            "format": "float",
            "description": "BM25权重",
            "default": 0.3
          },
          "vectorWeight": {
            "type": "number",
            "format": "float",
            "description": "向量权重",
            "default": 0.7
          },
          "numCandidates": {
            "type": "integer",
            "description": "kNN候选数量",
            "default": 100
          },
          "filters": {
            "type": "object",
            "description": "过滤条件"
          }
        }
      },
      "ApiResponseSearch": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object",
            "properties": {
              "query": {
                "type": "string",
                "description": "查询文本"
              },
              "searchType": {
                "type": "string",
                "description": "检索类型"
              },
              "total": {
                "type": "integer",
                "description": "结果总数"
              },
              "results": {
                "type": "array",
                "description": "检索结果列表",
                "items": {
                  "$ref": "#/components/schemas/SearchResult"
                }
              },
              "elapsedTime": {
                "type": "integer",
                "format": "int64",
                "description": "耗时(毫秒)"
              }
            }
          },
          "timestamp": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "SearchResult": {
        "type": "object",
        "properties": {
          "chunkId": {
            "type": "string",
            "description": "切片ID"
          },
          "documentId": {
            "type": "string",
            "description": "文档ID"
          },
          "content": {
            "type": "string",
            "description": "内容"
          },
          "score": {
            "type": "number",
            "format": "float",
            "description": "综合得分"
          },
          "bm25Score": {
            "type": "number",
            "format": "float",
            "description": "BM25得分"
          },
          "vectorScore": {
            "type": "number",
            "format": "float",
            "description": "向量得分"
          },
          "rank": {
            "type": "integer",
            "description": "排名"
          }
        }
      },
      "ApiResponseError": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "错误码"
          },
          "message": {
            "type": "string",
            "description": "错误消息"
          },
          "timestamp": {
            "type": "integer",
            "format": "int64",
            "description": "时间戳"
          }
        }
      }
    }
  }
}